+++ controls-core/src/jvmMain/kotlin/space/kscience/controls/spec/propertyReflection.kt
@@ -1,0 +1,8 @@
+package space.kscience.controls.spec
+
+import space.kscience.controls.api.PropertyDescriptor
+import kotlin.reflect.KProperty
+
+internal fun PropertyDescriptor.fromSpec(property: KProperty<*>){
+    property.annotations
+}